move identity verification logic into its own module
